+++ tests/tests/wgpu-gpu/external_image_copy.rs
